requirements.txt
# Automated Student Report Generator - Dependencies

pandas==2.3.3
reportlab==4.4.9
python-dotenv==1.2.1
